Write expect output to a temp file to ensure atomic file updates
With the cron process to update the data file and the munin fetch process
running the plugins operating independently, occasionally the plugins will be
run while the file is being updated but not fully written yet which can cause
failures.

Writing the updated results from the router to a separate file and then renaming
it guarantees the file will always be complete.