Initial commit of new Perl plugin
Mostly the same as the existing plugins in fuctionality. Major difference is
that the data fetch is done inside the plugin (with file locking to prevent
multiple instances from clobbering the shared state) so the cron job is no
longer required.

Some small bugfixes:
The sh plugins for atm1 and ptm1 would return counts of packets even if the
corresponding interface was not active.
For the vdsl_attenuation mode, the sh plugin was not returning the value for the
D3 band.